+++ slim_cli/template/main.py
@@ -1,5 +1,6 @@
 from app import app
 import config
+import permissions.roles_apply
 
 if __name__ == '__main__':
     import model._models