+++ slim_cli/template/model/__init__.py
@@ -35,12 +35,25 @@ class BaseModel(peewee.Model):
         return cls.select().where(cls._meta.primary_key == value).exists()
 
 
+class INETField(peewee.TextField):
+    field_type = 'inet'
+
+
+class CITextField(peewee.TextField):
+    field_type = 'CITEXT'
+
+
+class SerialField(peewee.IntegerField):
+    field_type = 'SERIAL'
+
+
 # asyncpg 配置
 asyncpg_conn = None
 
 
 def asyncpg_init(db_uri):
     import asyncpg
+
     async def create_conn():
         global asyncpg_conn
         asyncpg_conn = await asyncpg.connect(db_uri)