+++ slim_cli/template/model/_models.py
@@ -1,5 +1,7 @@
 from model import db
 from model.example import Example
+from model.user import User
+from model.user_token import UserToken
 
 db.connect()
-db.create_tables([Example], safe=True)
+db.create_tables([Example, User, UserToken], safe=True)