+++ slim_cli/template/view/__init__.py
@@ -3,5 +3,5 @@ import locale
 
 class ValidateForm(Form):
     class Meta:
-        # locales = ['zh_CN']
-        locales = [locale.getdefaultlocale()[0]]
+        locales = ['zh_CN']
+        # locales = [locale.getdefaultlocale()[0]]