+++ slim_cli/template/view/index.py
@@ -1,3 +1,4 @@
+from aiohttp import web
 from slim.base.view import BaseView
 from slim.retcode import RETCODE
 from app import app
@@ -8,7 +9,7 @@ class TestBaseView(BaseView):
     @app.route.interface('GET')
     async def info(self):
         self.finish(RETCODE.SUCCESS, {
-            'retcode': list(RETCODE.items()),
+            'retcode': RETCODE.to_dict(),
             'retinfo': RETCODE.txt_cn,
         })
 
@@ -16,3 +17,8 @@ class TestBaseView(BaseView):
     async def hello(self):
         data = await self.post_data()
         self.finish(RETCODE.SUCCESS, 'Hi, %s' % data.get('name', 'visitor'))
+
+
+@app.route('/api/aiohttp_request')
+async def hello(request):
+    return web.Response(text='hello')