+++ changelog.md
@@ -1,15 +1,99 @@
+#### 0.4.11 update 2019.09.09
+
+* Added: SQL Operator `contains` for ArrayField
+
+* Changed: Decorator `cooldown`'s parameter `interval_value_or_func` now accept async function
+
+
+
+#### 0.4.10 update 2019.09.08
+
+* Added: Decorator `must_be_role` 
+
+* Fixed: Decorator `cooldown` not work
+
+
+#### 0.4.9 update 2019.08.28
+
+* Added: Peewee PostgreSQL ArrayField supported
+
+* Added: `BaseView.method`
+
+* Added: `AbstractSQLView.current_interface`
+
+
+#### 0.4.8 update 2019.08.27
+
+* Changed: Modified return value of `view.params` and `view.post()` to `MultiDict`
+
+* Added: Decorator `require_role`
+
+
+#### 0.4.7 update 2019.08.25
+
+* Changed: template update
+
+
+#### 0.4.6 update 2019.08.25
+
+* Changed: template update
+
+* Changed(break): Rename decorator `auth_role` to `request_role`
+
+* Fixed: Correct return value of `view.params` and `view.post()` to `MultiDictProxy`
+
+
+#### 0.4.5 update 2019.08.24
+
+* Added: `msg` parameter for view.finish
+
+* Changed(break): Changed define of `Ability`
+
+* Changed: Updated cli template
+
+* Changed: Cli generate project with random port
+
+
+#### 0.4.4 update 2019.08.22
+
+* Added: decorator `auth_role` and decorator generator `get_cooldown_decorator`
+
+* Added: Type hints for UserViewMixin
+
+* Changed(break): Rewrite `BaseUserViewMixin`.
+
+* Changed(break): Python 3.6 required
+
+
+#### 0.4.3 update 2019.07.23
+
+* Changed: Renamed `XXConverter` to `XXParser`
+
+* Changed: Renamed `EMPTY_PERMISSION` to `NO_PERMISSION`
+
+
+#### 0.4.2 update 2019.06.12
+
+* Added: `LIST_PAGE_SIZE_CLIENT_LIMIT` for view
+
+* Changed(break): list api does not return `NOT_FOUND` anymore
+
+* Changed(break): Removed `app.route.get` `app.route.post` `app.route.head` and so on
+
+* Fixed: Can not insert record with empty data
+
+
 #### 0.4.1 update 2019.01.25
 
 * Fixed: compatible with peewee [3.8.2](https://github.com/coleifer/peewee/releases/tag/3.8.2).
 
 * Fixed: slim-cli not work
 
-
 #### 0.4.0 update 2019.01.24
 
 * Added: `view.temp_storage` for save values during the request
 
-* Added: `app.permission` option, it can be set to `ALL_PERMISSION`, `EMPTY_PERMISSION` or a `Permissions` object
+* Added: `app.permission` option, it can be set to `ALL_PERMISSION`, `NO_PERMISSION` or a `Permissions` object
 
 * Added: **`view.current_user` throw a Exception when current View not inherited from `BaseUserViewMixin`**